requirements.txt
crewai
langchain
langchain-openai
pydantic
python-dotenv
